+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.12.0
 
+- Binary name changed from `avail-light` to `avail-light-client`
 - Core changes are enumerated in [`core` CHANGELOG](../core/CHANGELOG.md) under the version 1.12.0
 
 ## [1.11.1](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.11.1) - 2024-07-23